+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - refactor `shared_set_tpm_path_constant` function
 - move all instructions to `docs/` dir
 - add `bin/install_plugins` cli executable script
+- improved test runner function
 
 ### v2.0.0, 2015-07-07
 - enable overriding default key bindings